invkinematics function

+++ d2c_robot_libs/include/d2c_robot_libs/robot_node.hpp
@@ -37,7 +37,7 @@ class D2cControl
         
         ros::Subscriber object_position_subscriber;
         ros::Subscriber control_keyboard_subscriber;
-        std::vector<float> target_joint_position;
+        std::vector<double> target_joint_position;
         int control_command_; // 0: Initialize 1: Pick up , 2: Pull down, 3: Stretch Arm, 4:Pull Arm
         float something;
         ServingCommand serving_command;
